+++ kajiApp_blazor/Components/View/Component/HomeComponent/WorkList.razor
@@ -1,4 +1,5 @@
-﻿@using kajiApp_blazor.Components.ViewModel
+﻿@using kajiApp_blazor.Components.DatabaseContext.HomeDBC
+@using kajiApp_blazor.Components.ViewModel
 @implements IDisposable
 @inject kajiappDBContext DbContext
 @using kajiApp_blazor.Components.Entity;
@@ -70,8 +71,8 @@
         {
             isLoading = true; // ローディング開始
             await InvokeAsync(StateHasChanged); // UI更新
-            var _homeDatabaseContext = new HomeDatabaseContext(DbContext);
-            worklist = await _homeDatabaseContext.GetWorksAsync();
+            var workListShow = new WorkListShow(DbContext);
+            worklist = await workListShow.GetWorksAsync();
         }
         catch (Exception ex)
         {
