+++ kajiApp_blazor/kajiApp_blazor.csproj
@@ -34,4 +34,8 @@
     <Folder Include="退避\" />
   </ItemGroup>
 
+  <ItemGroup>
+    <None Update="..\.dockerignore" Link="docker退避\.dockerignore" />
+  </ItemGroup>
+
 </Project>
